remove the workspace apps launcher from the tab strip
